requirements.txt
bs4
langchain
langchain-community
langchain-openai
python-dotenv
weaviate-client